Use buildNoticesURL instead of NoticesURL
We don't need to set up or store the noticesURL during construction
since we get this information at the time the error occurs with the
buildNoticesURL function instead. This ensures we always get the
freshest error host to use from the RemoteConfig or when that isn't
available just fallback to the default error host of `api.airbrake.io`.