Update minimum PHP to 8, upgrade dev dependencies
Update to 8.0 to be able to run with PHP 8 and installed with composer
on PHP 8, convert tests from old PHPUnit format to PHPUnit 8. Choosing
PHPUnit 8 as an intermediate step that has enough changes, but also has
some deprecations that will be removed in future PHP versions.

The only change to the library is to add a property that formerly was
implicit and to move the constant AIRBRAKE_NOTIFIER_VERSION from
Notifier.php to a separate file that will be autoloaded. This removes
the dependency on "accidentally" autoloading the constant when loading
the Notifier class. In the future we might think about making it a
class constant.

Update php_codesniffer to a version that's compatible with PHP 8