Fix specs, add src volume to docker-compose
With this specs go back to green.
A key step was setting the correct error_reporting level, at the time of
shutdown is was 22519 and that had no value in phpinfo(), Our own
notifier would set this as an E_UNKNOWN that doesn't trigger an error to
be sent.

Also adds an additional volume to the tests docker compose service so
changes in the `src` directory will be picked up on the next test run.